Implement enhancement described in issue #27
Update App.Views.VacationRequestsList so that it is possible to cancel
any vacation with status other than 'cancelled'.
Also, search filter is enabled. This functionality is provided
by the plugin out of box. For details on search filter see
http://bootstrap-table.wenzhixin.net.cn/documentation/#table-options

The search filter is supposed to replace previous filter mechanism
implementation.