Implement setting vacation to inprogress status
Vacations with status set to 'accepted' can be set to 'inprogress'
by vacation owner. This functionality is needed to indicate that
vacation is being in use.

Refactor templates for operations
All the buttons that are rendered in the column 'Operations',
in vacations table, are placed in a set of simple templates that
represent particular button each.

Add start action to RoR router
Add the action method as member of vacation_request resource.

Add authorization policy

Add template for 'Start' button
The button represents process of setting 'inprogress' status.

Add handler for 'Start' action

Cover all the changes with tests

Add some refactoring