Add available days to Pending Approval Requests
Fixes #47
This solution leads to changes in handling of the list
of available vacations.
As a result, from now on, the list of available vacations
for all the users can be fetched from '/available_vacations',
and the list of available vacations for particular user can be fetched
from '/users/<user_id>/available_vacations'.
Now, the form for creating new vacation requests updates the number
of available vacations after any vacation from the list is used.

Update RoR router with a new route

Update BB router
The router now propagates additional collections to views.
Functionality for updating the vacation request details,
the 'vacation_request_details', is removed.

Update the following RoR components according to changes:
  - UsersController
  - AvailableVacationsController

Update the following BB components according to changes:
  - App.Models.ApprovalRequest
  - App.Views.ApprovalRequests
  - App.Views.VacationRequestForm
  - App.Views.VacationRequests
  - App.Views.VacationRequestsList

Update some tests according to changes

Add a link to Holidays page
The http://www.timeanddate.com/holidays/ukraine is provided
as a useful link to pick up to date information.