Add BB Users page
The main idea is to provide users with the admin role with
the Users page.
The page represents the list of users with a Bootstrap Table.
The table provides controls for the following actions:
  - adding new user,
  - updating user info,
  - inviting user,
  - deleting user.
A modal dialog is provided for adding user and updating user info.
When user accepts an invitation, they get properly calculated
available vacations of all types. No manual work is needed here anymore.

Team model gets the following improvements:
  - minimum team name length is lowered to 3 letters,
  - maximum team name length is increased to 80 letters.

The following application components are updated:
  - Backbone.js, to 1.2.3
  - Bootstrap Table, to 1.9.1