Update Team policy
The main idea is to grant access to the Teams page to users with 'admin'
role only.

New RSpec matcher is added to work with URI fragments which can be
fetched from Capybara page object.
The URI fragment is that very part of URI that is used by BB router.
For more details on URI fragment, see the following documents:
  https://en.wikipedia.org/wiki/Fragment_identifier
  http://ruby-doc.org/stdlib-2.3.0/libdoc/uri/rdoc/URI.html